Homework 2 Task 1
NumberOfOccurences is a function that takes a number argument and returns 2 to the power of
that number.
estimationOfPi is also a function that takes a number argument (N-total) and returns the Pi
estimation.
MainQuestion1 utilizes both functions to plot two graphs that showcases the relationship 
between Pi estimation and N-total.