Homework 2 Task 2 Tol is a function that takes the number "n" as an argument and returns the value of 2 to the power of negative n.
estimationOfPiUsingTol is a function that takes Tol as an argument and uses a while loop to randomly generates a point that will determine whether it's in a circle or not. Pi_estimate is updated upon the completion of each loop until the while loop is satisfied and returns that Pi_estimate.

MainQuestion2 utilizes both functions to plot the time it takes to compute a Pi_estimation for a certain Tol against Tol.